fix: memberId에 따른 장르와 영화 매핑 수정

+++ src/main/java/com/capstone/cinemate/Movie/service/MovieService.java
@@ -132,9 +132,15 @@ public class MovieService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(IllegalArgumentException::new);
 
+        boolean hasExistingMemberMovies = !memberMovieRepository.findByMemberId(memberId).isEmpty();
+        boolean hasExistingGenreMembers = !genreMemberRepository.findByMemberId(memberId).isEmpty();
+
+        if (hasExistingMemberMovies || hasExistingGenreMembers) {
+            return;
+        }
+
         memberMovieRepository.deleteByMemberId(memberId);
         genreMemberRepository.deleteByMemberId(memberId);
-
 
         if (movieIds.size() > 6) {
             throw new CustomException(ErrorCode.BAD_REQUEST);
